Install .apps in ~/Applications if folder exists.
If the ~/Applications does not exist, it will automatically fallback to
/Applications.

More generically, any list passed in for the :prefix option will be
evaluated in-order, using the first directory that exists.

BaseDepRunner#app_dir has also been changed to only use the :prefix
option instead of hardcoding the local path (since this meta is the only
one that seems to use it and my new implementation will still search both
Application folders).