v0.9.0 - The old-style, deprecated 'pkg'/'gem'/etc methods to define deps on templates were removed. Lots of monkey patches were removed, delegate was removed, and there are no helpers mixed into the global scope anymore. All of this is to make babushka far more self-contained.
There's more work needed in this area. The main problem is that a bunch of core classes like Array and String are still being monkey patched. That's a tough one because it's really useful.

There were also some bugfixes, including one to escape 302 URLs when following them, from Saulius Grigaliunas.