v0.10.1 - Renamed Parameter#default to #suggest, and added a new #default.
Parameter#suggest behaves like the old :default #var option, i.e. it's the
suggested value when the user is prompted for the value of an unset parameter.

Parameter#default is now what the name implies - it's a default value for that
parameter that will be used without prompting the user if the parameter's value
is requested when it's unset. Setting the parameter to a value (on the
commandline, or when requiring the dep) overrides the default.