v0.10.2 - OK, let's all just pretend 0.10.1 never happened. Parameter#default behaves like #var's :default always has.
To do what I was trying to with 0.10.1, I've added Parameter#default!. Unset
parameters will return its value in preference to that of #default, and without
prompting, unless the user passes an explicit value from the command line, or
with #with when requiring the dep. So it behaves just like #default, except that
the user won't be prompted when the parameter is unset.

I was wrong to change the behaviour of Parameter#default in 0.10.1. I wouldn't
normally even consider a bait-and-switch like that, but since Parameter is brand
new, I made an exception. That was a mistake, though: it was still a surprising
change.

Anyway, this new design is better, I think :)