v0.10.3 - Fixed dep requirements with args on non-core sources. The problem was that the dep was being looked up eagerly, at the point the require happened, which can't work. The dep lookup is delayed now, to the point the dep is used by its parent in Dep#process_deps, where the lookup is done in the correct way.
When a dep is looked up, babushka needs to know which sources it should search
in. A dep that's required without a namespace is looked for first within the
same source as its parent, and then in the standard sources (i.e. the core set
that ship with babushka, and in standard locations like ~/.babushka/deps).

At the point of require, though, there's no way to tell what the current source
is without passing it or the current dep explicitly each time. When a dep was
required with args, using "name".with(args), babushka was trying to look up
eagerly in this way, and since the parent source wasn't supplied, it wasn't
finding deps required by their parent within the same source when they were
required with arguments.

(Deps required without args, i.e. as a string without a call to String#with,
are stored as-is, and so the lookup was delayed like it should be.)

In this update, I've redesigned "name".with(args) requirements to instead
return a Dep::Requirement, which is a little struct storing the dep's name and
the arguments it was required with, for later lookup. Nothing is actually
searched for now until the dep is required in Dep#process_deps, which is the
point it's invoked.

(Dep#process_deps passes the current source along to SourcePool#dep_for, so the
search knows which source to look for non-namespaced deps in.)