v0.10.5 - Fixed parameter re-wrapping, so they don't nest within each other. (Hey, that's the name of the show!) Also, fixed the way exceptions are handled and logged about during dep runs.
There are two types of exception that can be raised during a run - actual
errors that might indicate babushka bugs, and exceptions raised by babushka
when invalid deps are encountered (e.g. a dep defined against a nonexistent
template, or a dep required with the wrong number of arguments).

The exception handling code is much cleaner now (it's pretty much all within
Dep#process_this_dep), and the logging is more consistent and useful.