v0.10.10 - A fix for the "git is , which isn't >= 1.7.7" style logging.
Previously, it wasn't possible to tell if pieces of `cmd --version` output didn't
match because the version wasn't met, or because they weren't versions at all.

Now versions must contain a digit. This lets us make a good guess about what
parts of `cmd --version` are version numbers, so we can distinguish versions that
aren't matched with non-version bits.

(There is often numeric output in `cmd --version` that's not a version, like a
build date. However, I've never seen an example where it comes before the version
number, and we match on the first.)