make app deps with space in their name work
see https://github.com/benhoskings/babushka/issues/176 for details

Makes "Google Chrome.app" dep work. Was failing due to there being a
space in the name, which was funking out the version string handling.

Added a .valid? class level method to VersionStr to test if a given
string is in fact a valid version. This is called by VersionOf to check
the string before passing it as a name/version.

If it doesn't contain a valid version, just pass it as a name only
without a version to the VersionOf constructor.