Remove macports.
I don't use it anymore (I haven't for a while; homebrew is a much better packager for an OS X dev box). But that's not very important -- babushka supports other package managers like yum and pacman that I've never used.

The main reason is that there's no way to actually use Macports via babushka without changing the source -- MacportsHelper will never be selected otherwise. So unless there's a good reason to change that and make package managers configurable per-system (which I'm loathe to do because config options suck), it's not very useful having this code in place.