Add a special source name: "common" -> benhoskings/common-babushka-deps.
This source will be used for very widely applicable deps, for things like
deploying, remote running, and so on. The exception is there to allow the
"common" repo to live in my github account; otherwise, I'd have to manage
a separate github account for it (and find a good name for the repo whose
github account is still available).