Remove ruby1.8-dev from 'ruby', which usually won't be installed here.
The 'ruby' dep provides ruby and irb, which is met even if 'ruby1.8-dev' isn't installed. Since 'ruby' has to be present for us to be running this dep, the only way 'ruby1.8-dev' can be installed is if 'irb' is missing, and 'ruby1.8-dev' is pulled in along with it. That's surprising and best handled in a separate dep as required.