Don't try to avoid double-updating in 'apt source'; it backfires.
This logic was an attempt at efficiency, but it can cause the update
to be skipped.

When 'apt source' is called from 'keyed apt source', its updating is
skipped; however, if 'keyed apt source' is already met (i.e. the key
was present, but the source wasn't), then no update will happen and
the subsequent installation will fail because the packages are
missing.