Distinguish debian releases by major version.
The previous behaviour was incorrect: different minor versions of
debian aren't separate releases, but different versions of the same
release. https://www.debian.org/releases/

This obviates the need to add new SystemDefinitions entry per version,
as was addressed in PRs #308, #309 and so on (thanks Schnittchen).
Those are only required for new releases so babushka knows how to
refer to them, as with other OSes.

closes gh-313