Revert "Use PathHelpers.cd to wrap the chdir logic and avoid two exec calls."
This introduced a bug when --debug is in use.

Calling PathHelpers.cd from within Babushka::Open3's fork{} meant its
debugging printed not to babushka's STDOUT as expected, but to that
of the subprocess, which was then collected by Babushka::Shell as
output of the invoked command.

I caught this when I noticed that `babushka babushka --debug` behaved
differently to `babushka babushka`, which clearly should never happen.
In that case, the bug was causing Babushka::GitRepo.repo_for to
return the expected path (".git") with PathHelper's debugging output
prepended (needless to say, a nonexistent path):

"\e[90min dir /usr/local/babushka (/usr/local/babushka) {\e[m\n.git"

The real issue is that PathHelpers.cd is a user-facing helper, not a
low-level one, and so I shouldn't have used it in Babushka::Open3.

This reverts commit 1bb037502f487efa50bbad0c5374ef3111904635.