Fix remote babushka colouring logic.
This was incorrectly colouring when STDIN was a tty. That was wrong
for two reasons. Firstly, it's STDOUT that matters, since that's where
we're writing. And then secondly, the remote babushka's STDOUT will
never be a terminal; it's ssh.

The correct logic is for the remote to use the same colour setting as
the local babushka that's running it, whatever it's set to. Whether
that setting was autodetected from the terminal or forced with a
commandline option is irrelevant.

(Aside: we don't need to propogate the setting when colour is off.
Except when enabled with an explicit --colour, the remote will always
disable colour because of its non-tty STDOUT.)