Add `lbaction` to the list of non-descendable file extensions
These are LaunchBar extensions, distributed as Zip archives with an `.lbaction`, with a single top-level directory when extracted, also with an `.lbaction` extension. In order to make this top-level directory easy to move in one go, we don’t want to descend into it (e.g. exactly how Babushka also treats `tmbundle` directories).