Handle week -> month -> year thresholds nicely.
To do this, the divisor and threshold have to be different. A month is
30 days, but that leaves 28 and 29 days reporting "4 weeks" before we
jump to "1 month". Similarly, a year is 365 days, but that leaves us
reporting 360-364 days (after twelve 30-day months) as "12 months"
before we jump to "1 year". Instead, it's better to jump to the higher
threshold slightly early -- 28 days and 360 days are virtually one month
and one year respectively, anyway.