v0.19.0 - Remote babushkas over ssh; many fixes.
This version includes new helpers for easily running babushka (and other
shell commands) on a remote machine over ssh. I built the feature some
time ago but wanted to use it in anger for a while before I merged it.
Check the Babushka::SSH rdocs for a full explanation, but the gist is
that you can do things like this from within a dep:

    ssh('user@host').shell('uptime')
    ssh('user@host').babushka('postgresql')

It's a dumb implementation that doesn't do any connection reuse, etc.
That's partly because it's more predictable this way because it's just
doing what you'd otherwise be typing at a terminal. But more
importantly, the idea is to simply invoke a remote babushka to do all
the work locally on that box, rather than issue a series of remote
calls, because that's how you get truly good performance. As such the
overhead of a new connection per command shouldn't matter.

This release also includes lots of little bugfixes, including the
following:

* Fix colour enabling/disabling for local and remote babushkas,
  depending on output device, etc

* Properly recognise homebrew installations that weren't set up by
  babushka

* More reliable debian version detection, thanks to a report from Victor
  Klos

* When multiple deps with the given name are found across different
  default sources, use the first of them and issue a warning, instead of
  failing.

* Support weeks/months/years in friendly time text; previously it only
  went as far as days

* Bump the default git version to the most recent (at time of commit,
  v2.5.3)